libs/waivern-rulesets/pyproject.toml
[project]
name = "waivern-rulesets"
version = "0.1.0"
description = "Compliance rulesets for Waivern Compliance Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "waivern-core",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.2",
    "ruff>=0.11.12",
]

# Ruleset auto-discovery via entry points
[project.entry-points."waivern.rulesets"]
personal_data_indicator = "waivern_rulesets.personal_data_indicator:PersonalDataIndicatorRuleset"
processing_purposes = "waivern_rulesets.processing_purposes:ProcessingPurposesRuleset"
data_collection = "waivern_rulesets.data_collection:DataCollectionRuleset"
service_integrations = "waivern_rulesets.service_integrations:ServiceIntegrationsRuleset"
data_subject_indicator = "waivern_rulesets.data_subject_indicator:DataSubjectIndicatorRuleset"
gdpr_personal_data_classification = "waivern_rulesets.gdpr_personal_data_classification:GDPRPersonalDataClassificationRuleset"
gdpr_data_subject_classification = "waivern_rulesets.gdpr_data_subject_classification:GDPRDataSubjectClassificationRuleset"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# CRITICAL: Required for dev-mode installation
[tool.hatch.build]
dev-mode-dirs = ["src"]

# Include YAML ruleset data files and type marker in distribution
include = [
    "src/waivern_rulesets/**/*.py",
    "src/waivern_rulesets/data/**/*.yaml",
    "src/waivern_rulesets/py.typed",
]

[tool.hatch.build.targets.wheel]
packages = ["src/waivern_rulesets"]

[tool.basedpyright]
typeCheckingMode = "strict"
reportImplicitOverride = "error"

[tool.ruff]
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
    "ANN",    # flake8-annotations
    "B",      # flake8-bugbear
    "D",      # pydocstyle
    "F",      # pyflakes
    "I",      # isort
    "PL",     # pylint
    "RUF100", # unused-noqa-directive
    "S",      # bandit
    "UP",     # pyupgrade
]
ignore = [
    "D203",    # incorrect-blank-line-before-class (conflicts with D211)
    "D213",    # multi-line-summary-second-line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Use of assert detected - standard in pytest
    "PLR2004", # Magic value used in comparison - acceptable in tests
]
"src/waivern_rulesets/testing.py" = [
    "S101",    # Use of assert - required for test contract validation
    "PLR2004", # Magic value used in comparison - acceptable in test utilities
]
